Move project saving to server to helper, allow external save action
